update_2

+++ spicesapp/models.py
@@ -51,7 +51,7 @@ class Product(models.Model):
     product_name = models.CharField(max_length=100)
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    quantity = models.PositiveIntegerField()
+    quantity = models.DecimalField(max_digits=10, decimal_places=2)
     image = models.ImageField(upload_to='product_images', blank=True, null=True)
     description = models.TextField()
     product_code = models.CharField(max_length=20)
